librbd: moved IO operation tracker to io namespace
