+++ src/librbd/io/AioCompletion.h
@@ -10,8 +10,8 @@
 #include "include/utime.h"
 #include "include/rbd/librbd.hpp"
 
-#include "librbd/AsyncOperation.h"
 #include "librbd/ImageCtx.h"
+#include "librbd/io/AsyncOperation.h"
 #include "librbd/io/ReadResult.h"
 #include "librbd/io/Types.h"
