+++ src/librbd/io/CopyupRequest.h
@@ -4,11 +4,11 @@
 #ifndef CEPH_LIBRBD_IO_COPYUP_REQUEST_H
 #define CEPH_LIBRBD_IO_COPYUP_REQUEST_H
 
-#include "librbd/AsyncOperation.h"
 #include "include/int_types.h"
 #include "include/rados/librados.hpp"
 #include "include/buffer.h"
 #include "common/zipkin_trace.h"
+#include "librbd/io/AsyncOperation.h"
 #include "librbd/io/Types.h"
 
 #include <string>
